+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
+- BC-589 - BC-653 - Change auto deployment to a reusable workflow
+
 ## [26.16.0] - 2021-11-09
 
 ## [26.15.0] - 2021-11-02
@@ -103,7 +105,6 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
-- BC-589 - BC-653 - Change auto deployment to a reusable workflow
 - BC-68 - e2e run dependent on build image job
 - BC-234 - replaced hpi by datport in nbc imprint
 - BC-81 - removes old lern-store mode env var