+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
-- BC-589, BC-653 - Change auto deployment to a reusable workflow
+- BC-589 - BC-653 - change auto deployment to a reusable workflow
 
 ## [26.16.0] - 2021-11-09